% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqmongo.R
\name{buildJQterm}
\alias{buildJQterm}
\title{Build a single query function.}
\usage{
buildJQterm(name, value)
}
\arguments{
\item{name}{character name of the referenced field}

\item{value}{vector named collection of possible values.}
}
\value{
character scalar giving JSON expression
}
\description{
Build a single query function.
}
\details{
This is mostly an internal function, but may be of some use.
}
\examples{
buildJQterm("uid","Fred")
buildJQterm("uid",c("Phred","Fred"))
buildJQterm("time",Sys.time())
buildJQterm("num",1:4)
buildJQterm("num",c(gt=7))
buildJQterm("num",c(lt=7))
buildJQterm("num",c(gte=7))
buildJQterm("num",c(lte=7))
buildJQterm("num",c(ne=7))
buildJQterm("num",c(eq=7))
buildJQterm("num",c(gt=2,lt=7))
buildJQterm("count",c(nin=1,2:4))
buildJQterm("count",c("in"=1,2:4))
buildJQterm("count",c(ne=1,ne=5))
}
