% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\docType{class}
\name{JSONDB-class}
\alias{JSONDB-class}
\alias{mdbCRUD}
\alias{CRUD}
\title{Class which supports the mdbCRUD methods.}
\description{
Class which supports the mdbCRUD methods.
}
\details{
The CRUD (Create, Read, Update and Delete) are the basic set of
operators for manipulating a database.  The \code{mongo} package defines
a number of operators (mostly named \code{mdbXXX}) which calls the
corresponding CRUD operations.  The \code{JSONDB} class is intended for
any class that supports these operations.

The following operations are supported:
\itemize{
\item \code{\link[=mdbAvailable]{mdbAvailable()}} -- Returns logical value.  If false, CRUD operations
will basically be no-ops.
\item \code{\link[=mdbAggregate]{mdbAggregate()}} -- Runs an aggregation pipeline
\item \code{\link[=mdbCount]{mdbCount()}} -- Counts records matching query
\item \code{\link[=mdbDisconnect]{mdbDisconnect()}} -- Drops connection to database (will be
reconnected on next operation).
\item \code{\link[=mdbDistinct]{mdbDistinct()}} -- Lists unique values of a field.
\item \code{\link[=mdbDrop]{mdbDrop()}} -- Drops the collection from the database.  This is a
fast way to clear a database.
\item \code{\link[=mdbImport]{mdbImport()}},\code{\link[=mdbExport]{mdbExport()}} -- Imports/Exports documents
into/from a collection from a file (or connection).
\item \code{\link[=mdbFind]{mdbFind()}} -- Finds documents matching query and returns result
as a \code{data.frame}.
\item \code{\link[=mdbIndex]{mdbIndex()}} -- Adds or removes an index from a collection.
\item \code{\link[=mdbInfo]{mdbInfo()}} -- Returns info about a collection.
\item \code{\link[=mdbInsert]{mdbInsert()}} -- Adds one or more documents into a collections.
Works with both \code{data.frame} (one document per row) and JSON
character vectors (one document per element).
\item \code{\link[=mdbIterate]{mdbIterate()}}/\code{\link[=mdbFindL]{mdbFindL()}} -- Finds documents matching query and returns
these as an iterator/list.
\item \code{\link[=mdbMapreduce]{mdbMapreduce()}} -- Executes a mapreduce operation using
javascript map and reduce oerations.
\item \code{\link[=mdbRemove]{mdbRemove()}} -- Removes matching documents from a collection.
\item \code{\link[=mdbRename]{mdbRename()}} -- Renames a collection.
\item \code{\link[=mdbReplace]{mdbReplace()}}/\code{\link[=mdbUpsert]{mdbUpsert()}} -- Replaces a document in a
collection.
\item \code{\link[=mdbRun]{mdbRun()}} -- Runs a Mongo command.
\item \code{\link[=mdbUpdate]{mdbUpdate()}} -- Modifies records in a database.
\item \code{\link[=showDatabases]{showDatabases()}} -- Lists available databases.
\item \code{\link[=showCollections]{showCollections()}} -- Lists collections in a database.
}
}
