% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqmongo.R
\name{load_example}
\alias{load_example}
\title{Load example Event class}
\usage{
load_example()
}
\value{
invisible details about package
}
\description{
This function loads the example "Event" class; needed for examples.
}
\examples{
load_example()
fred1  ## sample data item.
## The source file
system.file("examples","Event.R",package="mongo")
}
