% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqmongo.R
\name{makeDBuri}
\alias{makeDBuri}
\title{Creates the URI needed to connect to a mongo database.}
\usage{
makeDBuri(
  username = "",
  password = "",
  host = "localhost",
  port = "",
  protocol = "mongodb"
)
}
\arguments{
\item{username}{The name of the database user (login credential), or an
empty string if no username is required.}

\item{password}{The name of the database password (login credential), or an
empty string if no password is required.}

\item{host}{The name or IP address of the system hosting the database.}

\item{port}{The port to be used for connections.  Note that the port for a
default configuration of mongo is 27018.  This can be left blank to use the
default port.}

\item{protocol}{A character scalar giving the protocol to use when
connecting, e.g., \dQuote{mongodb}.}
}
\value{
A character string giving the database URI which can be passed to the
\code{\link[mongolite]{mongo}} function to create a database collection
handle.

Note that the password is stored in clear text, so appropriate care should
be taken with the result of this function.
}
\description{
This function formats the universal record indicator (URI) for connecting to
a Mongo database.  It is mostly a utility function for formatting the
string.
}
\examples{

makeDBuri()
makeDBuri(user="admin",password="secret")
makeDBuri(user="admin")
makeDBuri(host="example.com",port=12345)

}
\seealso{
\code{\link{MongoDB}}, \code{\link[mongolite]{mongo}}

This is an input argument to a number of other classes which use mongo
connections.
}
\author{
Russell Almond
}
\keyword{database}
\keyword{interface}
