% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbAvailable}
\alias{mdbAvailable}
\alias{mdbAvailable,MongoDB-method}
\alias{mdbAvailable,mongo-method}
\title{Is the collection available for writing.}
\usage{
mdbAvailable(db)

\S4method{mdbAvailable}{MongoDB}(db)

\S4method{mdbAvailable}{mongo}(db)
}
\arguments{
\item{db}{(or MongoDB mongo) -- Reference to collection}
}
\value{
logical value.  If false, there is no active connection and CRUD operations will be no-ops.
}
\description{
Returns \code{FALSE} if the connection to the database is not available, so the CRUD operations (\link{mdbCRUD}).
will not be executed.
}
\note{
When using the \code{mongolite::\link[mongolite]{mongo}} collection reference
operations are not skipped.
}
