% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbCount}
\alias{mdbCount}
\alias{mdbCount,MongoDB-method}
\alias{mdbCount,mongo-method}
\title{Counts the number of records matching Query.}
\usage{
mdbCount(db, query = "{}")

\S4method{mdbCount}{MongoDB}(db, query = "{}")

\S4method{mdbCount}{mongo}(db, query = "{}")
}
\arguments{
\item{db}{MongoDB or mongo -- Reference to the collection}

\item{query}{character -- JSON expression giving the query. See
\verb{\link{buildJQuery}}.}
}
\value{
integer The number of records found (or \code{NA} if \code{noMongo = TRUE})
}
\description{
Counts the number of records matching Query.
}
\details{
The \code{query} argument is a partial match for the records (in JSON format) which is
essentially a partial match for the object.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
mdbCount(irisdb)
mdbCount(irisdb,'{"Species":"setosa"}')
mdbCount(irisdb,buildJQuery(Sepal.Width=c(lt=3),Petal.Width=c(gt=.3,lt=1.8)))
}
\seealso{
\verb{\link[mongolite]\{mongo\}}, \verb{\link{buildJQuery}}
\url{https://www.mongodb.com/docs/manual/reference/command/count/}
}
