% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbDisconnect}
\alias{mdbDisconnect}
\alias{mdbDisconnect,MongoDB-method}
\alias{mdbDisconnect,mongo-method}
\title{Disconnects connection to database.}
\usage{
mdbDisconnect(db, gc = TRUE)

\S4method{mdbDisconnect}{MongoDB}(db, gc = TRUE)

\S4method{mdbDisconnect}{mongo}(db, gc = TRUE)
}
\arguments{
\item{db}{MongoDB or mongo -- The database connection to drop.}

\item{gc}{logical -- Should the garbage collection be run.}
}
\value{
status message
}
\description{
Disconnects connection to database.
}
\details{
While this closes the connection, the \code{MongoDB} object retains the
information needed to re-open it.  It will be reopened on the next
call.
}
\examples{
## Setting noMongo=TRUE, so we don't actually run this.
testDB <- MongoDB("test", noMongo=!interactive())
mdbDisconnect(testDB)
}
\seealso{
[\link[mongolite]{mongo}]
}
