% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbDistinct}
\alias{mdbDistinct}
\alias{mdbDistinct,MongoDB-method}
\alias{mdbDistinct,mongo-method}
\title{Find the distinct values of a particular field}
\usage{
mdbDistinct(db, key, query = "{}")

\S4method{mdbDistinct}{MongoDB}(db, key, query = "{}")

\S4method{mdbDistinct}{mongo}(db, key, query = "{}")
}
\arguments{
\item{db}{(or MongoDB mongo) -- Reference to database collection}

\item{key}{character -- field to extract}

\item{query}{character -- JSON expression indicating subcollection.
See `\link{buildJQuery}`.}
}
\value{
list of values
}
\description{
Find the distinct values of a particular field
}
\details{
Finds the unique values of the field specified by `key`.  If
`query` is supplied, then search is restricted to records
satisfying query.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
mdbDistinct(irisdb,"Species")
}
\seealso{
[\link[mongolite]{mongo}]
\url{https://www.mongodb.com/docs/manual/reference/command/distinct/}
}
