% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbDrop}
\alias{mdbDrop}
\alias{mdbDrop,MongoDB-method}
\alias{mdbDrop,mongo-method}
\title{Drops the database collection}
\usage{
mdbDrop(db)

\S4method{mdbDrop}{MongoDB}(db)

\S4method{mdbDrop}{mongo}(db)
}
\arguments{
\item{db}{(or MongoDB mongo) -- Reference to collection to drop}
}
\value{
miniprint object giving status
}
\description{
Drops the database collection
}
\details{
Dropping the collection and then inserting values is an easy way to
reset the collection contents, so is a common idiom.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
}
\seealso{
[\link[mongolite]{mongo}]
\url{https://www.mongodb.com/docs/manual/reference/command/drop/}
}
