% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbInfo}
\alias{mdbInfo}
\alias{mdbInfo,MongoDB-method}
\alias{mdbInfo,mongo-method}
\title{Get Information about the collection}
\usage{
mdbInfo(db)

\S4method{mdbInfo}{MongoDB}(db)

\S4method{mdbInfo}{mongo}(db)
}
\arguments{
\item{db}{(or MongoDB mongo) The collection of interest.}
}
\value{
Object of class \code{miniprint} giving information about the collection.
}
\description{
Get Information about the collection
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
mdbInfo(irisdb)
}
\seealso{
[\link[mongolite]{mongo}]
}
