% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbMapreduce}
\alias{mdbMapreduce}
\alias{mdbMapreduce,MongoDB-method}
\alias{mdbMapreduce,mongo-method}
\title{Applies a summary operation to a collection}
\usage{
mdbMapreduce(db, map, reduce, query = "{}", sort = "{}", limit = 0)

\S4method{mdbMapreduce}{MongoDB}(db, map, reduce, query = "{}", sort = "{}", limit = 0)

\S4method{mdbMapreduce}{mongo}(db, map, reduce, query = "{}", sort = "{}", limit = 0)
}
\arguments{
\item{db}{(or MongoDB mong) -- The collection to operate on.}

\item{map}{character -- A javascript function to apply to each document.}

\item{reduce}{character -- A javascript function to summarize the result}

\item{query}{character -- A JSON query to slect part of the
collection.  See `\link{buildJQuery}` and `\link{mdbFind}`.}

\item{sort}{charcter -- JSON object giving sorting order for result set (see `\link{mdbFind}`).}

\item{limit}{integer -- maximum number of records to process}
}
\value{
data frame with results
}
\description{
Runs a map-reduce operation in the database side.  The `map` and
`reduce` functions are expressed as javascript methods.
}
\details{
The Mongo database manual suggests that aggregation piplelines have
better performance than map-reduce.  Starting in Mongo 5.0
map-reduce is depricated.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
histdata <- mdbMapreduce(irisdb,
  map= "function (){emit(Math.floor(this.Petal_Length*5)/5, 1)}",
  reduce="function (id,counts){return Array.sum(counts)}"
)
if (any(!is.na(histdata))) {
names(histdata) <- c("Petal.length","count")
}
head(histdata)
histdata1 <- mdbAggregate(irisdb,
 '[{"$set": {
    "ptlround": {
      "$divide":[
           {"$floor": {
            "$multiply": ["$Petal_Length", 5]
            }}, 5]}}},
   {"$group": {
     "_id": "$ptlround",
     "count": {"$sum":1}
   }}
]'
)
if (!is.null(histdata1)) {
names(histdata1) <- c("Petal.length","count")
}
head(histdata1)
}
\seealso{
[\link[mongolite]{mongo}] [\link{mdbAggregate}]
\url{https://www.mongodb.com/docs/manual/core/map-reduce/}
}
