% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbRemove}
\alias{mdbRemove}
\alias{mdbRemove,MongoDB-method}
\alias{mdbRemove,mongo-method}
\title{Remove selected objects from collection}
\usage{
mdbRemove(db, query = "{}", just_one = FALSE)

\S4method{mdbRemove}{MongoDB}(db, query = "{}", just_one = FALSE)

\S4method{mdbRemove}{mongo}(db, query = "{}", just_one = FALSE)
}
\arguments{
\item{db}{(or MongoDB mongo) -- Collection affected.}

\item{query}{character -- Mongo query expressed as JSON object.
See `\link{buildJQuery}` and `\link{mdbFind}`.}

\item{just_one}{logical -- If true, only the first matching record
is removed.}
}
\value{
miniprint Information about the results.
}
\description{
Query selects a subset of the collection to remove.  Note for
removing everything, `\link{mdbDrop}` is faster.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbDrop(irisdb)
mdbInsert(irisdb,iris)
mdbCount(irisdb)
mdbRemove(irisdb,'{"Species":"setosa"}')
mdbCount(irisdb)
}
\seealso{
[\link[mongolite]{mongo}] [\link{mdbFind}]
[\link{mdbDrop}]
\url{https://www.mongodb.com/docs/manual/reference/command/delete/}
}
