% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbRename}
\alias{mdbRename}
\alias{mdbRename,MongoDB-method}
\alias{mdbRename,mongo-method}
\title{Renames collection or moves it to a new database}
\usage{
mdbRename(mdb, name, db = NULL)

\S4method{mdbRename}{MongoDB}(mdb, name, db = NULL)

\S4method{mdbRename}{mongo}(mdb, name, db = NULL)
}
\arguments{
\item{mdb}{(or MongoDB mongo) -- Reference to collection to move}

\item{name}{character -- new name for collection}

\item{db}{character -- new database for collection.}
}
\value{
miniprint Status Message
}
\description{
Using the `name` argument simply renames the collection.  Using the
`db` argument copies the collection to a new database.
}
\examples{
mdbDrop(MongoDB("FisherIrises",noMongo=!interactive()))
irisdb <- MongoDB("iris",noMongo=!interactive())
showCollections(irisdb)
mdbRename(irisdb,"FisherIrises")
showCollections(irisdb)
}
\seealso{
[\link[mongolite]{mongo}]
\url{https://www.mongodb.com/docs/manual/reference/command/renameCollection/}
}
