% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbRun}
\alias{mdbRun}
\alias{mdbRun,MongoDB-method}
\alias{mdbRun,mongo-method}
\title{Runs a Mongo command on the collection}
\usage{
mdbRun(db, command = "{\\"ping\\":1}", simplify = TRUE)

\S4method{mdbRun}{MongoDB}(db, command = "{\\"ping\\":1}", simplify = TRUE)

\S4method{mdbRun}{mongo}(db, command = "{\\"ping\\":1}", simplify = TRUE)
}
\arguments{
\item{db}{(or MongoDB mongo) Reference to the collection.}

\item{command}{character JSON document providing the command.}

\item{simplify}{logical If true, the output structure is simplified.}
}
\value{
list containing returned value.
}
\description{
Runs a Mongo command on the collection
}
\details{
A command is a JSON document.  See the Mongo reference manual for
the supported commands (these will vary quite a lot by the version
of the database).  Note that some commands only run against the
"admin" database.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
mdbRun(irisdb,'{"collStats":"iris"}')
}
\seealso{
\verb{\link[mongolite]\{mongo\}}
\url{https://www.mongodb.com/docs/manual/reference/command/}
}
