% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{mdbUpdate}
\alias{mdbUpdate}
\alias{mdbUpdate,MongoDB-method}
\alias{mdbUpdate,mongo-method}
\title{Modify document(s) in a collection}
\usage{
mdbUpdate(
  db,
  query,
  update = "{\\"$set\\":{}}",
  filters = NULL,
  upsert = FALSE,
  multiple = FALSE
)

\S4method{mdbUpdate}{MongoDB}(
  db,
  query,
  update = "{\\"$set\\":{}}",
  filters = NULL,
  upsert = FALSE,
  multiple = FALSE
)

\S4method{mdbUpdate}{mongo}(
  db,
  query,
  update = "{\\"$set\\":{}}",
  filters = NULL,
  upsert = FALSE,
  multiple = FALSE
)
}
\arguments{
\item{db}{(or MongoDB mongo) -- The collection to modify.}

\item{query}{character -- A JSON document identifying the document(s)
to modify.}

\item{update}{character -- A JSON document identifying the changes
to make to a document.}

\item{filters}{character -- A JSON document which controls which documents in the collection
gets updated.  (See the "Mongolite User Manual").}

\item{upsert}{logical -- If true and the query returns no results,
then insert the document instead.}

\item{multiple}{logical -- If true, all documents matching `query`
are affected; if false, only the first one.}
}
\value{
A list with the returned object.
}
\description{
The `query` field identifies a number of documents.  The `update`
is a set of instructions for changing the documents.  The
`\link{mdbReplace}` function replaces the document instead of
modifying it.
}
\details{
### Queries

The rules here match those for `\link{mdbFind}`.

### Update Documents

This is a special document which describes how to modify the
document.  There are a number of commands described in the Mongo
documentation, of which the two most useful are `$set` and
`$unset`.

The `$set` command takes as argument a JSON object giving field
value pairs.  If the field exists it value will be changed, if it
doesn't exist a new field will be created.  For example `'{"$set":
{"processed":false}}'` will set the `processed` field to false,
additing it if necessary.

The `$unset` command removes a field from a document.  For example,
`'{"$unset":{"processed":0}}'` removes the process field.  The
value after the field name is ignored.

For multiple complex changes, consider using
`\link{mdbAggregate} for more complex changes.
}
\examples{
 mdb <- MongoDB("testthis","test","mongodb://localhost",noMongo=!interactive())
 mdbDrop(mdb)
 mdbInsert(mdb,c('{"name":"Fred", "gender":"M"}',
                 '{"name":"George", "gender":"M"}'))
mdbFind(mdb)
mdbUpdate(mdb,'{"name":"Fred"}', '{"$set":{"gender":"F"}}')
mdbFind(mdb)
}
\seealso{
[\link[mongolite]{mongo}] [\link{mdbReplace}]
[\link{mdbFind}] [\link{mdbAggregate}]
}
