% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\docType{class}
\name{MongoDB-class}
\alias{MongoDB-class}
\alias{MongoDB}
\title{MongoDB -- R6 class wrapping a connection to a Mongo database collection.}
\usage{
MongoDB(
  collection = "test",
  db = "test",
  url = "mongodb://localhost",
  verbose = FALSE,
  noMongo = FALSE,
  options = mongolite::ssl_options()
)
}
\arguments{
\item{collection}{character -- name of collection}

\item{db}{character -- name of database}

\item{url}{character -- URI for mongo connection (see [makeDBuri()])}

\item{verbose}{logical -- Should operate in verbose mode.}

\item{noMongo}{logical -- If true, then no connection to Mongo database will be made,
and CRUD operations will become no-ops.}

\item{options}{-- SSL options for connections, see [mongolite::ssl_options()].}
}
\description{
MongoDB -- R6 class wrapping a connection to a Mongo database collection.
}
\section{Fields}{

\describe{
\item{\code{mongoObj}}{ANY -- This is the actual `\link[mongolite]{mongo}` object or `NULL` if it has not been initialized yet.}

\item{\code{uri}}{character -- URI for the mongo connection.}

\item{\code{dbname}}{character -- The name of the mongo database.}

\item{\code{colname}}{character -- The name of the mongo collection}

\item{\code{noMongo}}{logical -- If `TRUE`, then the various methods will turn into NO-OPS.
This allows a class which contains a reference to a Mongo database to ignore the database calls when
there is no database to connect to.}

\item{\code{verbose}}{logical -- This field is passed on to the `\link[mongolite]{mongo}` call.}

\item{\code{options}}{ANY -- This field is passed on to the  `\link[mongolite]{mongo}` call.  It is used to store additional SSL connection information.

Including a \link[mongolite]{mongo} object in an R6 (Reference) class presents a potential race condition.
The prototype class is built at package load time, however, calling the `\link[mongolite]{mongo}` may not work
at this time.  The `MongoDB` class works around this by capturing the arguments to the `mongo` call, and then
creating the actual database connection when the database is first accessed.  The database should always be accessed
through the `$db()` method which builds the database if needed.}
}}


\note{
Many of the examples use `MongoDB(...,noMongo=!interactive())`.
This means the dummy mechanism will be used during package checking
(where Mongo may or may not be available in the development
environment), but running the examples from the help files will
make the connections (and will generate an error if Mongo is not
installed).
}
\section{Methods}{


These two R6 methods are mostly for internal use.

* `$db()` -- Returns an object of class `\link[mongolite]{mongo}` which is
the actual databse, or `NULL` if `uri==""` or `noMongo==TRUE`.
This will open the database if it has not yet been created,
or otherwise return the stashed database.
* `$noMongo()` logical -- Returns the contents of the `noMongo` field.  Used to suppress actual mongo calls
* `$resetDB()` -- This sets the internal `mongoObj` field to `NULL` so
the connection will be rebuilt on the next call to `$db()`.  This is probably
useful to call when restoring an R Session from

For the most part, the following S4 generic functions provide the normal database operation on the collection:
`\link{mdbAggregate}`, `\link{mdbCount}`, `\link{mdbDisconnect}`, `\link{mdbDrop}`,
`\link{mdbExport}`, `\link{mdbFind}`, `\link{mdbImport}`, `\link{mdbIndex}`,
`\link{mdbInsert}`, `\link{mdbIterate}`, `\link{mdbMapreduce}`,
`\link{mdbRemove}`, `\link{mdbRename}`, `\link{mdbReplace}`,
`\link{mdbRun}`, `\link{mdbUpdate}`.
}

\examples{
mdp <- MongoDB("test","test","mongodb://localhost")
\dontrun{
# This will generate an error if mongo doesn't exist.
mdbCount(mdp,'{}')
}
nullmdp <- MongoDB(noMongo=TRUE)
mdbCount(nullmdp)
# This will return `NA`.
}
\seealso{
[\link[mongolite]{mongo}]
More extensive documentation on most of the `mdbXXX` functions can be found at the Mongo API documentation web site.
\url{https://www.mongodb.com/docs/manual/reference/command/}
}
