% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\docType{class}
\name{MongoRec-class}
\alias{MongoRec-class}
\alias{MongoRec}
\title{Class "MongoRec".}
\usage{
MongoRec(..., m_id = NA_character_)
}
\arguments{
\item{...}{Other arguments (pass through for initialization method).}

\item{m_id}{character Mongo identifier.  Use \code{NA_character_} for unsaved objects.}
}
\value{
The constructor \code{MongoRec} returns an object of class \code{MongoRec}.

The \code{m_id} method returns a character scalar (with name \code{oid}) which contains the mongo identifer.
}
\description{
This is a lightweight class meant to be extended.
It contains a
single field for a Mongo identifier, which can be accessed using
the \code{m_id()} method.  It is meant to store something that is a
record in a Mongo collection, where \verb{_id} is the Mongo identifier.
}
\section{Functions}{
\itemize{
\item \code{MongoRec()}: Constructor for \code{MongoRec}

}}
\section{Slots}{

\describe{
\item{\code{_id}}{(character) The Mongo ID, \code{NA_character_} if not saved.}
}}

\section{Objects from the Class}{


Objects can be created by calls to the \code{MongoRec()} function.
}

\examples{

showClass("MongoRec")

}
\seealso{
\code{\link[=as.json]{as.json()}}
\code{\link[=buildObject]{buildObject()}}, \code{\link[=saveRec]{saveRec()}}, \code{\link[=getOneRec]{getOneRec()}}
}
\author{
Russell G. Almond
}
\keyword{classes}
