% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\name{parse.jlist}
\alias{parse.jlist}
\alias{parse.jlist,ANY,list-method}
\alias{buildObject}
\alias{parse.json}
\alias{parse.jlist,MongoRec,list-method}
\title{Construct an S4 object from a list of its slot values.}
\usage{
parse.jlist(class, rec)

\S4method{parse.jlist}{ANY,list}(class, rec)

buildObject(rec, class = rec$class)

parse.json(encoded, builder = buildObject)

\S4method{parse.jlist}{MongoRec,list}(class, rec)
}
\arguments{
\item{class}{-- A character string defining the class of the output object.
If the list has an element named `class`, that will be used.}

\item{rec}{-- A list which is the output of \code{\link[jsonlite]{fromJSON}}}

\item{encoded}{-- a character scalar giving the raw JSON object.}

\item{builder}{-- A function which will construct an object from a list of fields values.}
}
\value{
An S4 object of type `class`
}
\description{
The `parse.json` function uses the
\code{\link[jsonlite]{fromJSON}} function to turn the JSON into a list, which
is processed using the function \code{parse.jlist}
to massage the elements, and then passes it to the \code{new} function
to create a new object of type \code{class}.
}
\details{
The \code{parse.jlist} function is a helper function designed to do any massaging
necessary to unencode the slot values before the object is produced.  The function
\code{\link{ununboxer}} undoes the effect of \code{unboxer}, and the
function \code{\link{unparseData}} undoes the effect of \code{parseData}.
}
\section{Functions}{
\itemize{
\item \code{parse.jlist()}: This is the inner function for processing
the slots prior to object creation.  Generally, this is the method
that needs to be specialized.  See the `vignette("JSON for S4
Objects")`.

\item \code{parse.jlist(class = ANY, rec = list)}: Base case for callNextMethod; just returns
the slot list.

\item \code{buildObject()}: This method takes the jlist, cleans it with an appropriate
`parse.jlist` method and then tries to generate an object based on the class.

\item \code{parse.jlist(class = MongoRec, rec = list)}: Makes sure the `_id` field corresponds to
conventions, and inserts `NA` if it is missing.

}}
\examples{
\dontrun{
vignette("JSON for S4 Objects")
}
}
