% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\name{parsePOSIX}
\alias{parsePOSIX}
\title{Convert Mongo dates to POSIX}
\usage{
parsePOSIX(x)
}
\arguments{
\item{x}{-- Either a list of the form \code{list("$date"=1234)} or any other value
compatible with \verb{link[base]\{as.POSIXct\}}.}
}
\value{
the contents of \code{x} as a \code{POSIXct} object.
}
\description{
Converting a date to Mongo-flavored JSON produces
a numeric value (number of second seconds since Jan 1, 1970)
labeled with \verb{$date}.  This function takes the output of
\verb{\link[jsonlite]\{fromJSON\}} and converts it back to POSIX format.
}
\details{
If the date has been marked as scalar (using the \verb{\link{unboxer}} or
\verb{\link[jsonlite]\{unbox\}}, this function will strip the \code{scalar} flag.
}
\examples{
dt <- Sys.time()
dtj <- jsonlite::toJSON(unboxer(dt))
parsePOSIX(jsonlite::fromJSON(dtj,FALSE))
}
