% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\name{parseSimpleData}
\alias{parseSimpleData}
\title{parseSimpleData}
\usage{
parseSimpleData(messData)
}
\arguments{
\item{messData}{list output from \verb{\link[jsonlite]\{fromJSON\}}}
}
\value{
list, simplified.
}
\description{
Simple parser works with data which is mostly primitive R types (\code{numeric}, \code{integer}, \code{logical},
\code{character}).
}
\details{
The \verb{\link[jsonlite]\{fromJSON\}} method does not distinguish between arrays type character,
logical, integer or numeric.  This function finds lists of a single atomic type and
replaces them with the corresponding \code{vector}.
}
\examples{
parseSimpleData(list(chars=list("a","b","c"),nums=list(2.3,3.4,4.5),
  ints=list(1,2,3), logic=list(TRUE,FALSE)))
}
