% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{showCollections}
\alias{showCollections}
\alias{showCollections,mongo-method}
\alias{showCollections,MongoDB-method}
\alias{showCollections,NULL-method}
\title{Shows collections in the current database.}
\usage{
showCollections(
  db = NULL,
  dbname = "test",
  uri = "mongodb://localhost",
  options = mongolite::ssl_options()
)

\S4method{showCollections}{mongo}(
  db = NULL,
  dbname = "test",
  uri = "mongodb://localhost",
  options = mongolite::ssl_options()
)

\S4method{showCollections}{MongoDB}(
  db = NULL,
  dbname = "test",
  uri = "mongodb://localhost",
  options = mongolite::ssl_options()
)

\S4method{showCollections}{`NULL`}(
  db = NULL,
  dbname = "test",
  uri = "mongodb://localhost",
  options = mongolite::ssl_options()
)
}
\arguments{
\item{db}{(or MongoDB mongo NULL) -- Connection to target database.
If `NULL` a new connection is made using the specified information.}

\item{dbname}{character -- name for new collection}

\item{uri}{character -- URI for database connections.}

\item{options}{list -- SSL options for an SSL connection}
}
\value{
character vector of database names
}
\description{
Shows collections in the current database.
}
\details{
Shows all of the collections which are in the referenced database.
If the `db` argument is a `MongoDB` or `mongolite::mongo` object,
the current database is used.  If `db` is `NULL`, then a new
connection is created with the information.
}
\examples{
irisdb <- MongoDB("iris",noMongo=!interactive())
showCollections(irisdb)
}
\seealso{
[\link[mongolite]{mongo}]
}
