% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\docType{class}
\name{MongoRec-class}
\alias{MongoRec-class}
\alias{MongoRec}
\title{Class "MongoRec".}
\usage{
MongoRec(..., m_id = NA_character_)
}
\arguments{
\item{...}{Other arguments (pass through for initialization method).}

\item{m_id}{character Mongo identifier.  Use `NA_character_` for unsaved objects.}
}
\value{
The constructor `MongoRec` returns an object of class `MongoRec`.

The `m_id` method returns a character scalar (with name `oid`) which contains the mongo identifer.
}
\description{
This is a lightweight class meant to be extended.
It contains a
single field for a Mongo identifier, which can be accessed using
the `m_id()` method.  It is meant to store something that is a
record in a Mongo collection, where `_id` is the Mongo identifier.
}
\section{Functions}{
\itemize{
\item \code{MongoRec()}: Constructor for `MongoRec`

}}
\section{Slots}{

\describe{
\item{\code{_id}}{(character) The Mongo ID, `NA_character_` if not saved.}
}}

\section{Objects from the Class}{


Objects can be created by calls to the `MongoRec()` function.
}

\examples{

showClass("MongoRec")

}
\seealso{
[as.json()]
[buildObject()], [saveRec()], [getOneRec()]
}
\author{
Russell G. Almond
}
\keyword{classes}
